+++ CHANGELOG.md
@@ -2,15 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 
+---
+## [1.0.1] - 2022-06-08
+### Added
+- Add ABE key generation: master key generation and user decryption key generation
+- Add attributes rotation
+### Changed
+- Depends on `abe_gpsw` v0.8.0 and `cover_crypt` v3.0.1
+### Fixed
+### Removed
 
 ---
 ## [1.0.0] - 2022-05-17
 ### Added
 - Implementation of ABE-GPSW/AES hybrid encryption/decryption using `abe_gpsw` v0.6.9
-- Implementation of ABE-CovCrypt/AES hybrid encryption/decryption using `cover_crypt` v1.0.2
+- Implementation of ABE-CovCrypt/AES hybrid encryption/decryption using `cover_crypt` v2.0.0
 - Implementation of HKDF HMAC-SHA256
 ### Changed
-- Depends on `abe_gpsw` v0.6.9
 ### Fixed
 ### Removed
 ---